requirements.txt
discord.py==2.7.1
flask==3.1.3
requests==2.33.1
